requirements.txt
pydantic>=2.8.2
numpy>1.21
torch>1.13
typing_extensions>4.6
wandb>0.15
rlgym-learn==0.1.7